# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TrueStage DataBridge API - AI-powered data ingestion and mapping platform for credit union data onboarding
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /ingestion-jobs:
    get:
      operationId: listIngestionJobs
      tags: [ingestion-jobs]
      summary: List ingestion jobs
      parameters:
        - name: creditUnionId
          in: query
          required: false
          schema:
            type: ["integer", "null"]
        - name: status
          in: query
          required: false
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of ingestion jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/IngestionJob"

    post:
      operationId: createIngestionJob
      tags: [ingestion-jobs]
      summary: Create a new ingestion job (file upload)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateIngestionJobBody"
      responses:
        "201":
          description: Created ingestion job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestionJob"
  /ingestion-jobs/{id}:
    get:
      operationId: getIngestionJob
      tags: [ingestion-jobs]
      summary: Get an ingestion job by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ingestion job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestionJob"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /ingestion-jobs/{id}/process:
    post:
      operationId: processIngestionJob
      tags: [ingestion-jobs]
      summary: Trigger AI mapping and processing of an ingestion job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Processing started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestionJob"
...
  /ingestion-jobs/{id}/records:
    get:
      operationId: listIngestionRecords
      tags: [processed-records]
      summary: List processed records for an ingestion job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: status
          in: query
          required: false
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProcessedRecord"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    IngestionJob:
      type: object
      properties:
        id:
          type: integer
        creditUnionId:
          type: integer
        creditUnionName:
          type: ["string", "null"]
        fileName:
          type: string
        fileType:
          type: string
        fileSize:
          type: ["integer", "null"]
        status:
          type: string
          enum: [uploaded, mapping, mapped, reviewing, processing, completed, failed, partial]
        totalRecords:
          type: ["integer", "null"]
        processedRecords:
          type: ["integer", "null"]
        acceptedRecords:
          type: ["integer", "null"]
        softErrorRecords:
          type: ["integer", "null"]
        hardErrorRecords:
          type: ["integer", "null"]
        mappingConfidenceAvg:
          type: ["number", "null"]
        errorSummary:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - creditUnionId
        - fileName
        - fileType
        - status
        - createdAt
        - updatedAt
    CreateIngestionJobBody:
      type: object
      properties:
        creditUnionId:
          type: integer
        fileName:
          type: string
        fileType:
          type: string
        fileSize:
          type: ["integer", "null"]
        fileContent:
          type: string
          description: Base64 encoded file content or raw text
      required:
        - creditUnionId
        - fileName
        - fileType
        - fileContent
...
    ProcessedRecord:
      type: object
      properties:
        id:
          type: integer
        ingestionJobId:
          type: integer
        rowNumber:
          type: integer
        status:
          type: string
          enum: [accepted, soft_error, hard_error]
        canonicalData:
          type: ["string", "null"]
          description: JSON of the mapped canonical fields
        rawData:
          type: ["string", "null"]
          description: JSON of the original raw fields
        errorDetails:
          type: ["string", "null"]
          description: JSON array of error details
        errorSeverity:
          type: ["string", "null"]
          enum: [soft, hard, null]
        createdAt:
          type: string
      required:
        - id
        - ingestionJobId
        - rowNumber
        - status
        - createdAt
...